refactor(agent_harness): split the service module by concern
domain/service.rs had grown to 1,500 lines of one type's impl blocks.
It is now a module directory: mod.rs keeps the service type, its
constructor and public entry points; queue.rs owns the per-session
command queue, its worker and cross-replica routing; open.rs owns the
three ways a session is created; deliver.rs owns composing and prompting;
lifecycle.rs owns control events, sandbox sizing, turn boundaries and
teardown. No behaviour changes; methods called across files are
pub(super).
